Table: split TableNG utils tests alongside their modules
Breaks the 2989-line utils.test.ts into six files, one per module created in the
previous commit, so a test file sits next to the code it covers. Each imports the
symbols it exercises from its module by direct path rather than through the
barrel, which keeps the tests honest about where each function lives.

The 35 describe blocks move verbatim, dedented one level now that the outer
`TableNG utils` wrapper is gone. The buildInspectValue snapshots move with them
to utils/__snapshots__/display.test.ts.snap, with the wrapper dropped from their
keys; the snapshot bodies are unchanged.

Suite totals are identical before and after: 819 tests and 22 snapshots across
the Table folder.
